+++ setup.cfg
@@ -20,6 +20,8 @@ ignore =
     W503
     # whitespace before :
     E203
+    # redefs
+    F811
 
 [isort]
 known_first_party=fsspec
